+++ src/core/transform/ReservedWords.ts
@@ -1,0 +1,8 @@
+export const reservedWords = new Set([
+  "break", "case", "catch", "class", "const", "continue", "debugger", "default", "delete",
+  "do", "else", "export", "extends", "false", "finally", "for", "function", "if", "import",
+  "in", "instanceof", "new", "null", "return", "super", "switch", "this", "throw", "true",
+  "try", "typeof", "var", "void", "while", "with", "yield", "let", "static", "enum", "await",
+  "implements", "package", "protected", "interface", "private", "public", "arguments", "eval",
+  "undefined", "of", "async", "get", "set"
+]);